// src/app/core/models/domain/venta.model.ts
import { Cliente } from './cliente.model';
import { FormaVenta } from './forma-venta.model';
import { VentaDetalle } from './venta-detalle.model';
import { Cobranza } from './cobranza.model';
import { MotivoAnulacion } from './motivo-anulacion.model';
import { Pedido } from './pedido.model';
import { Deposito } from './deposito.model';
import { Vendedor } from './vendedor.model';
import { ListaPrecio } from './lista-precio.model';
import { Canal } from './canal.model';
import { Terminales } from './terminales.model';
export interface Venta {
  codVenta: number;
  codEmpresaErp: string;
  codSucursalErp: string;
  codVendedorErp: string;
  listaPrecio: ListaPrecio;
  anulado: boolean;
  codEmpresa: number;
  codSucursal: number;
  estado: string;
  modoEntrega: string;
  fechaAnulacion: Date;
  fechaCreacion: Date;
  fechaVencimiento: Date;
  fechaVenta: string;
  fechaModificacion: string;
  porcDescuento: number;
  importeDescuento: number;
  importeIva5: number;
  importeIva10: number;
  importeIvaExenta: number;
  importeNeto: number;
  importeTotal: number;
  descuentoProducto: number;
  subTotal: number;
  totalKg: number;
  timbrado: string;
  inicioTimbrado: string;
  finTimbrado: string;
  codUsuarioAnulacion: number;
  nroComprobante: string;
  tipoComprobante: string;
  terminalVenta: Terminales;
  deposito: Deposito;
  codUsuarioCreacion: number;
  cliente: Cliente;
  pedido: Pedido;
  formaVenta: FormaVenta;
  detalle: VentaDetalle[];
  motivoAnulacion: MotivoAnulacion;
  esObsequio: boolean;
  editable: boolean;
  cobranza: Cobranza;
  vendedor: Vendedor;
  tipoVenta: string;
  canal: Canal;
  cupon: string;
  codReparto?: number;
  totalPuntos?: number;
  puntosObtenidos?: number;
}