// src/app/core/models/domain/cliente.model.ts
import { Sucursal } from './sucursal.model';
import { ListaPrecio } from './lista-precio.model';
import { MedioPago } from './medio-pago.model';
import { FormaVenta } from './forma-venta.model';
import { GrupoDescuento } from './grupo-descuento.model';
export interface Cliente {
  codCliente: number;
  codClienteErp: string;
  razonSocial: string;
  tipoDoc: string;
  docNro: string;
  codEmpresa: number;
  latitud: number;
  longitud: number;
  sucursalPref: Sucursal;
  listaPrecio: ListaPrecio;
  medioPagoPref: MedioPago;
  formaVentaPref: FormaVenta;
  grupoDescuento: GrupoDescuento;
  carnetGrupo: string;
  carnetVencimiento: string;
  categoria?: string;
  grupo?: string;
  zona?: string;
  direccion?: string;
  telefono?: string;
  email?: string;
  web?: string;
  catABC?: string;
  obs?: string;
  activo?: boolean;
  codeBarra?: boolean;
  diasCredito?: number;
  fechaCreacion?: string;
  fechaModificacion?: string;
  codUltimaVenta: number;
  excentoIva?: boolean;
  concatCodErpNombre?: string;
  concatCodNombre?: string;
  concatDocNombre?: string;
  predeterminado?: boolean;
  esPropietario?: boolean;
  empleado?: boolean;
  puntos?: number;
}